app_backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==3.14.0
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==8.4.0
Markdown==3.7
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.4
zipp==3.20.1
